+++ components/icon/button-icon/index.jsx
@@ -9,22 +9,79 @@ Neither the name of salesforce.com, inc. nor the names of its contributors may b
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 'AS IS' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-import React from 'react';
-import SLDSUtilityIcon from '../../utilities/utility-icon';
+// ### React
+import React, { PropTypes } from 'react';
+
+// ### classNames
+// [github.com/JedWatson/classnames](https://github.com/JedWatson/classnames)
+// This project uses `classnames`, "a simple javascript utility for conditionally
+// joining classNames together."
 import classNames from 'classnames';
 
-const displayName = 'ButtonIcon';
+// This component's `checkProps` which issues warnings to developers about properties when in development mode (similar to React's built in development tools)
+import checkProps from './check-props';
+
+// ### Children
+import SLDSUtilityIcon from '../../utilities/utility-icon';
+
+import { BUTTON_ICON } from '../../../utilities/constants';
+
+/**
+ * This is a non-interctive wrapper component for `UtilityIcon` that specifies button icon classes for an icon inside a `button` tag. Use of this component by itself is not recommended, but should be used as part of other components to obtain the correct styling for icons within buttons. This component only partially implements [Button Icons](http://www.lightningdesignsystem.com/components/button-icons). It does not return a `button` HTML tag. It only returns an icon for use within a button. Assistive text must also be rendered by the parent.
+ */
+const ButtonIcon = (props) => {
+	checkProps(BUTTON_ICON, props);
+
+	return (
+		<SLDSUtilityIcon
+			aria-hidden="true"
+			category={props.category}
+			className={classNames('slds-button__icon', {
+				[`slds-button__icon--${props.size}`]: props.size && props.size !== 'medium',
+				'slds-button__icon--inverse-hint': props.inverse && props.hint,
+				'slds-button__icon--hint': props.hint && !props.inverse,
+				[`slds-button__icon--${props.position}`]: props.position
+			},
+			props.className)}	// iconClassName has been deprecated
+			icon={props.icon}
+			name={props.name}
+		/>
+	);
+};
 
 const propTypes = {
-	assistiveText: React.PropTypes.string,
-	category: React.PropTypes.oneOf(['action', 'custom', 'doctype', 'standard', 'utility']).isRequired,
-	className: React.PropTypes.string,
-	hint: React.PropTypes.bool,
-	icon: React.PropTypes.object,
-	inverse: React.PropTypes.bool,
-	name: React.PropTypes.string,
-	position: React.PropTypes.oneOf(['left', 'right']),
-	size: React.PropTypes.oneOf(['x-small', 'small', 'medium', 'large'])
+	/**
+	 * Icon category from [lightningdesignsystem.com/icons/](https://www.lightningdesignsystem.com/icons/)
+	 */
+	category: PropTypes.oneOf(['action', 'custom', 'doctype', 'standard', 'utility']).isRequired,
+	/**
+	 * Associates an icon button with another element on the page by changes the color of the SVG. Please reference <a href="http://www.lightningdesignsystem.com/components/buttons/#hint">Lightning Design System Buttons > Hint</a>.
+	 */
+	hint: PropTypes.bool,
+	/**
+   * An SVG object to use instead of name / category, look in `design-system-react/icons` for examples
+   */
+	icon: PropTypes.object,
+	/**
+	 * Class names to be added to the SVG.
+	 */
+	className: PropTypes.oneOfType([PropTypes.array, PropTypes.object, PropTypes.string]),
+	/**
+	 * Setting `inverse` to true will switch the color of the icon: light to dark, dark to light.
+	 */
+	inverse: PropTypes.bool,
+	/**
+	 * Name of the icon. Visit <a href='http://www.lightningdesignsystem.com/resources/icons'>Lightning Design System Icons</a> to reference icon names.
+	 */
+	name: PropTypes.string,
+	/**
+	 * Adds additional spacing on the opposite side specified between button icon and the button label
+	 */
+	position: PropTypes.oneOf(['left', 'right']),
+	/**
+	 * Size of the icon. Visit [lightningdesignsystem.com/components/icons/#flavor-sizes](https://www.lightningdesignsystem.com/components/icons/#flavor-sizes)
+	 */
+	size: PropTypes.oneOf(['x-small', 'small', 'medium', 'large'])
 };
 
 const defaultProps = {
@@ -32,44 +89,7 @@ const defaultProps = {
 	size: 'medium'
 };
 
-class ButtonIcon extends React.Component {
-	constructor (props) {
-		super(props);
-		this.state = {};
-	}
-
-	getClassName () {
-		return classNames(this.props.className, 'slds-button__icon', {
-			[`slds-button__icon--${this.props.position}`]: this.props.position,
-			[`slds-button__icon--${this.props.size}`]: this.props.size && this.props.size !== 'medium',
-			'slds-button__icon--inverse-hint': this.props.inverse && this.props.hint,
-			'slds-button__icon--hint': this.props.hint && !this.props.inverse
-		});
-	}
-
-	render () {
-		let label = null;
-
-		if (this.props.assistiveText) {
-			label = <span className="slds-assistive-text">{this.props.assistiveText}</span>;
-		}
-
-		return (
-			<span>
-				{label}
-				<SLDSUtilityIcon
-					aria-hidden="true"
-					category={this.props.category}
-					className={this.getClassName()}
-					icon={this.props.icon}
-					name={this.props.name}
-				/>
-			</span>
-		);
-	}
-}
-
-ButtonIcon.displayName = displayName;
+ButtonIcon.displayName = BUTTON_ICON;
 ButtonIcon.propTypes = propTypes;
 ButtonIcon.defaultProps = defaultProps;