+++ stories/dropdown/index.jsx
@@ -6,6 +6,7 @@ import { storiesOf, action } from '@kadira/storybook';
 
 import { MENU_DROPDOWN } from '../../utilities/constants';
 import Dropdown from '../../components/menu-dropdown';
+import { DropdownNubbinPositions } from '../../components/menu-dropdown/menu-dropdown';
 import List from '../../components/menu-list/list';
 import Button from '../../components/button';
 import Trigger from '../../components/menu-dropdown/button-trigger';
@@ -27,6 +28,43 @@ const getDropdown = (props) => (
 	<Dropdown {...props} />
 );
 
+const getDropdownPositioned = (props) => {
+	const positionedDropdowns = [];
+	DropdownNubbinPositions.forEach((position) => {
+		positionedDropdowns.push(
+			<div className="slds-col slds-size--1-of-3" style={{ minHeight: '500px' }}>
+				<Dropdown
+					{...props}
+					forceOpen
+					nubbinPosition={position}
+				>
+					<Trigger>
+						<Button iconVariant="container" iconName="settings" label={position} />
+					</Trigger>
+				</Dropdown>
+			</div>
+		);
+	});
+
+	return (
+		<div>
+			<div className="slds-grid slds-wrap">
+				{positionedDropdowns}
+			</div>
+			<div className="slds-col" style={{ minHeight: '500px' }}>
+				<Dropdown
+					{...props}
+					nubbinPosition="top right"
+				>
+					<Trigger>
+						<Button iconVariant="container" iconName="settings" assistiveText="top right" />
+					</Trigger>
+				</Dropdown>
+			</div>
+		</div>
+	);
+};
+
 const getDropdownCustomTrigger = (props) => (
 	<Dropdown {...props} >
 		<Trigger>
@@ -35,21 +73,27 @@ const getDropdownCustomTrigger = (props) => (
 	</Dropdown>
 );
 
-const getDropdownCustomContent = (props) => (
-	<Dropdown {...props} >
-		<div id="custom-dropdown-menu-content">
-			<div className="slds-m-around--medium">
-				<div className="slds-tile slds-tile--board slds-m-horizontal--small">
-					<p className="tile__title slds-text-heading--small">Art Vandelay</p>
-					<div className="slds-tile__detail">
-						<p className="slds-truncate">
-							<a className="slds-m-right--medium" href="#">Settings</a>
-							<a href="#" >Log Out</a>
-						</p>
-					</div>
+/* eslint-disable react/prop-types */
+/* eslint-disable no-script-url */
+const DropdownCustomContent = (props) => (
+	<div id="custom-dropdown-menu-content">
+		<div className="slds-m-around--medium">
+			<div className="slds-tile slds-tile--board slds-m-horizontal--small">
+				<p className="tile__title slds-text-heading--small">Art Vandelay</p>
+				<div className="slds-tile__detail">
+					<p className="slds-truncate">
+						<a className="slds-m-right--medium" href="javascript:void(0)" onClick={props.onClick}>Settings</a>
+						<a href="javascript:void(0)" onClick={props.onClick}>Log Out</a>
+					</p>
 				</div>
 			</div>
 		</div>
+	</div>
+);
+
+const getDropdownCustomContent = (props) => (
+	<Dropdown {...props} >
+		<DropdownCustomContent />
 		<List options={[{ label: 'Custom Content Option' }, ...options]} />
 	</Dropdown>
 );
@@ -67,6 +111,13 @@ storiesOf(MENU_DROPDOWN, module)
 	.add('No Modal', () => getDropdown({
 		align: 'right',
 		label: 'Dropdown Click',
+		onSelect: (...rest) => {
+			action('Selected')(...rest);
+		},
+		modal: false,
+		options
+	}))
+	.add('No Modal w/ Nubbins', () => getDropdownPositioned({
 		onSelect: (...rest) => {
 			action('Selected')(...rest);
 		},