+++ stories/global-navigation-bar/index.jsx
@@ -26,7 +26,7 @@ const dropdownCollection = [{
 	label: 'Main action', value: '0', iconCategory: 'utility',
 	iconName: 'table', href: 'http://www.google.com'
 }, {
-	label: 'Menu Header', type: 'header'
+	label: 'Menu Header', type: 'header', divider: 'top'
 }, {
 	label: 'Menu Item One', value: '1', iconCategory: 'utility',
 	iconName: 'kanban', href: 'http://www.google.com'
@@ -39,6 +39,8 @@ const dropdownCollection = [{
 }, {
 	label: 'Menu Item Four', value: '4', iconCategory: 'utility',
 	iconName: 'side_list', href: 'http://www.google.com'
+}, {
+	type: 'divider'
 }, {
 	label: 'Menu Item Five', value: '5', iconCategory: 'utility',
 	iconName: 'side_list', href: 'http://www.google.com'
@@ -77,22 +79,23 @@ const getGlobalNavigationBar = (props, primaryRegionProps) => (
 			<GlobalNavigationBarDropdown
 				assistiveText="Open Menu Item 1"
 				id="primaryDropdown"
-				label="Menu Item 1"
+				label="Menu Item"
+				openOn={props.openOn || undefined}
 				onSelect={dropdownItemClicked('Dropdown Menu Item clicked')}
 				options={dropdownCollection}
 			/>
 			<GlobalNavigationBarLink
 				// will actually go to website
 				href="https://www.lightningdesignsystem.com/"
-				label="Menu Item 2"
+				label="Menu Item"
 			/>
 			<GlobalNavigationBarLink
 				active
-				label="Menu Item 3"
+				label="Menu Item"
 				onClick={linkClicked('Link clicked')}
 			/>
 			<GlobalNavigationBarLink
-				label="Menu Item 4"
+				label="Menu Item"
 				onClick={linkClicked('Link clicked')}
 			/>
 		</GlobalNavigationBarRegion>
@@ -300,6 +303,7 @@ storiesOf(GLOBAL_NAVIGATION_BAR, module)
 			propSets.noNav.props,
 			propSets.noNav.primaryRegionProps
 		)
-	);
+	)
+	.add('Hybrid Dropdown', () => getGlobalNavigationBar(propSets.hybrid.props, propSets.base.primaryRegionProps));
 
 module.exports = getGlobalNavigationBar;